페이지 목록 기능 구현

+++ src/main/java/com/example/demo/mybatis/commons/pager/Pager.java
@@ -13,31 +13,31 @@ import lombok.ToString;
 public class Pager {
 	//--- ### 변수 선언 영역.
 	// 페이지의 번호
-	private int pageNum = 1;
+	private long pageNum = 1;
 	// 게시글의 개수
-	private int totalBoard = 0;
+	private long totalBoard = 0;
 	// 페이지에 출력되는 게시글 개수 (n개씩 보기)
-	private int pageSize = 10;
+	private long pageSize = 10;
 	// 한 블럭에 출력되는 페이지 번호 개수
-	private int blockSize = 10;
+	private long blockSize = 10;
 	//---
 	// 전체 페이지 개수
-	private int totalPage;
+	private long totalPage;
 	// 요청 페이지에 출력되는 게시글의 시작 행번호
-	private int startRow;
+	private long startRow;
 	// 요청 페이지에 출력되는 게시글의 종료 행번호
-	private int endRow;
+	private long endRow;
 	// 블럭에 출력되는 시작 페이지 번호
-	private int startPage;
+	private long startPage;
 	// 블럭에 출력되는 종료 페이지 번호
-	private int endPage;
+	private long endPage;
 	// 이전 블럭에 출력되는 시작 페이지 번호
-	private int prevPage;
+	private long prevPage;
 	// 다음 블럭에 출력되는 시작 페이지 번호
-	private int nextPage;
+	private long nextPage;
 	//--- ### 생성자 선언 영역.
 	public Pager() {}
-	public Pager(int pageNum, int totalBoard, int pageSize, int blockSize) {
+	public Pager(long pageNum, long totalBoard, long pageSize, long blockSize) {
 		this.pageNum = pageNum;
 		this.totalBoard = totalBoard;
 		this.pageSize = pageSize;
@@ -48,7 +48,7 @@ public class Pager {
 	//--- ### 메서드 선언 영역.
 	public void calcPage() {
 		// 총 게시물 수에 페이지당 표시되는 게시물 수를 나눠 전체 페이지의 개수를 구합니다.
-		totalPage=(int)Math.ceil((double)totalBoard/pageSize);
+		totalPage=(long)Math.ceil((double)totalBoard/pageSize);
 		// 페이지 번호가 전체 페이지 수 보다 큰 경우 페이지 번호를 1로 설정
 		if(pageNum<=0 || pageNum>totalPage) {
 			pageNum=1;
